+++ js/data.js
@@ -1,0 +1,31 @@
+const RENT_TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const RENT_CHECKIN = ['12:00', '13:00', '14:00'];
+const RENT_CHECKOUT = ['12:00', '13:00', '14:00'];
+const RENT_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const RENT_FOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+const MIN_LAT = 35.65000;
+const MAX_LAT = 35.70000;
+const MIN_LNG = 139.70000;
+const MAX_LNG = 139.80000;
+const MAX_ADS = 10;
+const MAX_GUEST = 5;
+const MAX_ROOMS = 7;
+const MAX_PRICE = 22800;
+const ads = [];
+
+export {
+  RENT_TYPE,
+  RENT_CHECKIN,
+  RENT_CHECKOUT,
+  RENT_FEATURES,
+  RENT_FOTOS,
+  MIN_LAT,
+  MAX_LAT,
+  MIN_LNG,
+  MAX_LNG,
+  MAX_ADS,
+  MAX_GUEST,
+  MAX_ROOMS,
+  MAX_PRICE,
+  ads
+};
